+++ src/BookOperations.java
@@ -1,0 +1,5 @@
+public interface BookOperations {
+
+    void displayDetails();
+
+}
